+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project follows semantic versioning.
 
 ### Unreleased
 - [added] Support for parent-specific proc-macros.
+- [fixed] Support for const generics, associated lifetimes, nested generics.
 
 ### 1.1.3 (2025-10-01)
 - [changed] Bumped dependency versions